Update ArtifactRegistry.yaml
Each of the controls are categorized as `Protect`, though each of their comments mentions that once the control is deployed, it can `Detect` nefarious activity. Should the said controls be mapped to both a Protect and Detect categories? With the given description I'd recommend `Detect` only. Example: Compare T1068 with T1212.